Genericized the concept of conditions
Conditions previously exclusively owned an internal trigger string which
was matched against the the content of a visible text message.

Conditions now own a function (signature `func (ctx Context) bool` )
that returns true or false given a context.

Context are the a summary of the environment in which an event takes
place that could match a bot condition.