Voice connection can persist across voice queue
Bot checks its current voice connection before sending a voice payload.
Voice connection only disconnects / changes channel if a voice payload
indicates it should be dispatched to a different channel than the one
the bot is already in.

Disconnect when no voice payloads are waiting to be processed.

Added a new type of action "emojiReactionAction" for appending emoji
reactions to messages.

Added some conditions for reacting to messages related to heroes of the
storm.

Code reorganization.