Refactored conditions
Sacrificed the flexibility of conditions holding an arbitrary function
that determines when the condition is met.
Added method func (ctx *Context) Satisfies(c Condition) bool to
determine when a condition is met.
Conditions now hold descriptions of contexts that would meet them.
Empty values are treated as wildcards.  E.g. if a condition does not
specify a user, it can be satisfied by any user.
This effort brings conditions closer to the goal of querying conditions
as entities in a database.